+++ packages/agents-core/src/auth/permissions.ts
@@ -1,5 +1,6 @@
+import type { AccessControl } from 'better-auth/plugins/access';
 import { createAccessControl } from 'better-auth/plugins/access';
-import { adminAc, defaultStatements } from 'better-auth/plugins/organization/access';
+import { defaultStatements } from 'better-auth/plugins/organization/access';
 
 const statement = {
   ...defaultStatements,
@@ -16,7 +17,7 @@ const statement = {
   context_config: ['create', 'read', 'update', 'delete'],
 } as const;
 
-export const ac = createAccessControl(statement);
+export const ac = createAccessControl(statement) as AccessControl
 
 export const memberRole = ac.newRole({
   project: ['read'],
@@ -44,7 +45,6 @@ export const adminRole = ac.newRole({
   external_agent: ['create', 'read', 'update'],
   function: ['create', 'read', 'update'],
   context_config: ['create', 'read', 'update'],
-  ...adminAc.statements,
 });
 
 export const ownerRole = ac.newRole({
@@ -59,5 +59,4 @@ export const ownerRole = ac.newRole({
   external_agent: ['create', 'read', 'update', 'delete'],
   function: ['create', 'read', 'update', 'delete'],
   context_config: ['create', 'read', 'update', 'delete'],
-  ...adminAc.statements,
 });